expose span to M::assert_panic, and provide helper to turn that into CallerLocation
